requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
rich>=13.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
